requirements.txt
gitdb2==2.0.3
GitPython==2.1.8
smmap2==2.0.3
